+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added support for rendering 3D volume textures.
 
+### Changed
+* Changed xc3_wgpu_batch to render wismhd models to individual files.
+
 ### Fixed
 * Fixed an issue where `xc3_model::Animation::fcurves` would not include root motion in some cases.
 * Fixed an issue where xc3_wgpu would not properly render animations for skeletons with more than 256 bones.